Reorganize common/dto
