+++ packages/common/index.ts
@@ -1,1 +1,2 @@
 export * from './src/hello-world';
+export * from './src/dto';
