+++ packages/common/src/dto/availability-information.dto.ts
