+++ packages/common/src/dto/contact-information.dto.ts
@@ -1,18 +1,6 @@
 import { IsEmail, IsOptional, IsString, Length } from 'class-validator';
 
-export class PersonalInformationDTO {
-  @IsString()
-  @Length(2, 20)
-  firstName!: string;
-
-  @IsString()
-  @Length(2, 20)
-  lastName!: string;
-
-  @IsString()
-  @Length(6, 7)
-  postalCode!: string;
-
+export class ContactInformationDTO {
   @IsString()
   @Length(6, 12)
   primaryPhone!: string;
