+++ packages/common/src/dto/form.dto.ts
@@ -1,5 +1,6 @@
 import { IsIn, IsString, ValidateNested } from 'class-validator';
 import { AvailabilityDTO } from './availability-information.dto';
+import { ContactInformationDTO } from './contact-information.dto';
 import { PersonalInformationDTO } from './personal-information.dto';
 import { SkillInformationDTO } from './skill-information.dto';
 
@@ -7,6 +8,7 @@ export class FormDTO {
   @ValidateNested()
   payload!: {
     personalInformation: PersonalInformationDTO;
+    contactInformation: ContactInformationDTO;
     skillInformation: SkillInformationDTO;
     availabilityInformation: AvailabilityDTO;
   };
