+++ packages/common/src/dto/index.ts
@@ -1,0 +1,5 @@
+export * from './personal-information.dto';
+export * from './contact-information.dto';
+export * from './skill-information.dto';
+export * from './placement-preferences.dto';
+export * from './form.dto';
