+++ packages/common/src/dto/placement-preferences.dto.ts
